split tests better
